requirements.txt
apache-airflow==2.11.0
psycopg2-binary==2.9.10
